chore: split runner/dumper
