+++ rhoc/pkg/cmd/clusters/list/list.go
@@ -3,41 +3,21 @@ package list
 import (
 	"errors"
 	"fmt"
-	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/util/cmdutil"
-	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/util/dumper"
-	"net/http"
-	"strconv"
-	"time"
-
 	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/api/admin"
 	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/service"
-	"github.com/olekukonko/tablewriter"
+	"github.com/bf2fc6cc711aee1a0c2a/cos-tools/rhoc/pkg/util/cmdutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/dump"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
 	"github.com/spf13/cobra"
+	"net/http"
+	"strconv"
 )
 
 const (
 	CommandName  = "list"
 	CommandAlias = "ls"
 )
-
-type cluster struct {
-	ID         string
-	Owner      string
-	CreatedAt  time.Time
-	ModifiedAt time.Time
-	State      string
-}
-type clusterWide struct {
-	ID         string
-	Name       string
-	Owner      string
-	CreatedAt  time.Time
-	ModifiedAt time.Time
-	State      string
-}
 
 type options struct {
 	outputFormat string
@@ -148,48 +128,4 @@ func run(opts *options) error {
 	}
 
 	return nil
-}
-
-func dumpAsTable(f *factory.Factory, items admin.ConnectorClusterList, wide bool) {
-	r := make([]interface{}, 0, len(items.Items))
-
-	for i := range items.Items {
-		k := items.Items[i]
-
-		if wide {
-			r = append(r, clusterWide{
-				ID:         k.Id,
-				Name:       k.Name,
-				Owner:      k.Owner,
-				CreatedAt:  k.CreatedAt,
-				ModifiedAt: k.ModifiedAt,
-				State:      string(k.Status.State),
-			})
-		} else {
-			r = append(r, cluster{
-				ID:         k.Id,
-				Owner:      k.Owner,
-				CreatedAt:  k.CreatedAt,
-				ModifiedAt: k.ModifiedAt,
-				State:      string(k.Status.State),
-			})
-		}
-	}
-
-	t := dumper.NewTable(map[string]func(s string) tablewriter.Colors{
-		"state": statusCustomizer,
-	})
-
-	t.Dump(r, f.IOStreams.Out)
-}
-
-func statusCustomizer(s string) tablewriter.Colors {
-	switch s {
-	case "ready":
-		return tablewriter.Colors{tablewriter.Normal, tablewriter.FgHiGreenColor}
-	case "disconnected":
-		return tablewriter.Colors{tablewriter.Normal, tablewriter.FgHiBlueColor}
-	}
-
-	return tablewriter.Colors{}
 }
